\name{example_gammas}
\alias{example_gammas}
\docType{data}
\title{
Example Gamma Value
}
\description{
This is the list of gamma values used in the examples.
}
\usage{data(example_gammas)}
\format{
  The format is:
 chr "example_gammas"
}


\examples{
data(example_gammas)
## maybe str(example_gammas) ; plot(example_gammas) ...
}
\keyword{datasets}
