% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_export.R
\name{geo_export}
\alias{geo_export}
\title{Export geospatial and/or attribute data to a file}
\usage{
geo_export(x, path, filetype, na = NA, overwrite = FALSE)
}
\arguments{
\item{x}{Name of R data object to output. REQUIRED.}

\item{path}{Filepath specifying location and name of output file to create. REQUIRED.}

\item{filetype}{Type of output file to create.  Must be one of the following: "shp" (default),
"csv" (.csv comma separated value file), "sas" (.sas7bdat data file),
"excel" (.xlsx workbook file). REQUIRED.}

\item{na}{Character string to be used for missing values in the output file. Replaces all
empty values ("" and " "), and properly coded missing values (NA), with the specified
string. If no string is provided, default is NA. OPTIONAL.}

\item{overwrite}{If output dataset already exists at the path location, should R overwrite
(replace existing file) or not?  Default is to not overwrite an existing file. OPTIONAL.}
}
\value{

}
\description{
This function exports spatial and attribute data to file.  The output file type
can be a shapefile (.shp), comma separated value file (.csv), Excel workbook (.xlsx),
or a SAS data file (.sas7bdat).
}
\examples{
# simple export shp file ------------------------------------------------------------------
geo_export(x = my_dat,
 path = my_outdir,
 filetype = "shp",
 overwrite = TRUE)

# exporting when file already exists ------------------------------------------------------
geo_export(x = my_dat,
 path = my_outdir,
 filetype = "shp",
 overwrite = FALSE) # will not replace existing file

geo_export(x = my_dat,
 path = my_outdir,
 filetype = "shp",
 overwrite = TRUE) # replaces existing file

# specify new value ('unknown') to replace empty and missing values in output file --------
geo_export(x = my_dat,
 path = my_outdir,
 filetype = "shp",
 na = "unknown",
 overwrite = TRUE)
}
